+++ src/components/custom-field-input-multiple-choice/custom-field-input-multiple-choice.test.jsx
@@ -26,31 +26,73 @@ describe('<CustomFieldInputMultipleChoice>', () => {
   });
 
   describe('autocompleter popup', () => {
-    it('opens on click', () => {
-      render(<CustomFieldInputMultipleChoice {...requiredProps} />);
-      expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
-      userEvent.click(screen.getByLabelText('test label'));
-      expect(screen.queryByRole('listbox')).toBeInTheDocument();
+    describe('input', () => {
+      it('focuses on click', () => {
+        render(<CustomFieldInputMultipleChoice {...requiredProps} />);
+        expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+        userEvent.click(screen.getByText('test label'));
+        expect(screen.getByLabelText('test label', { selector: 'input' })).toHaveFocus();
+      });
+
+      it('filters visible choices on value', () => {
+        render(<CustomFieldInputMultipleChoice {...requiredProps} />);
+        expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+        userEvent.click(screen.getByText('test label'));
+        expect(screen.getByLabelText('test label', { selector: 'input' })).toHaveFocus();
+        expect(screen.getByText('Choice 1')).toBeInTheDocument();
+        expect(screen.getByText('Choice 2')).toBeInTheDocument();
+        userEvent.type(document.activeElement, '1');
+        expect(screen.getByText('Choice 1')).toBeInTheDocument();
+        expect(screen.queryByText('Choice 2')).not.toBeInTheDocument();
+      });
+
+      it('is cleared on selection', () => {
+        render(<CustomFieldInputMultipleChoice {...requiredProps} />);
+        userEvent.click(screen.getByText('test label'));
+        expect(screen.getByText('Choice 1')).toBeInTheDocument();
+        expect(screen.getByText('Choice 2')).toBeInTheDocument();
+        userEvent.type(document.activeElement, '1');
+        userEvent.click(screen.getByText('Choice 1'));
+        expect(screen.getByLabelText('test label', { selector: 'input' })).toHaveValue('');
+      });
     });
 
-    it('closes on escape key', () => {
-      render(<CustomFieldInputMultipleChoice {...requiredProps} />);
-      userEvent.click(screen.getByLabelText('test label'));
-      expect(screen.queryByRole('listbox')).toBeInTheDocument();
-      fireEvent.keyDown(document.activeElement, { key: 'Escape' });
-      expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
-    });
+    describe('popup', () => {
+      it('opens on click', () => {
+        render(<CustomFieldInputMultipleChoice {...requiredProps} />);
+        expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+        userEvent.click(screen.getByText('test label'));
+        expect(screen.queryByRole('listbox')).toBeInTheDocument();
+      });
 
-    it('does not open when there are no choices', () => {
-      render(<CustomFieldInputMultipleChoice {...requiredProps} choices={[]} />);
-      userEvent.click(screen.getByLabelText('test label'));
-      expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
-    });
+      it('opens on typing', async () => {
+        render(<CustomFieldInputMultipleChoice {...requiredProps} />);
+        userEvent.tab();
+        expect(screen.getByLabelText('test label', { selector: 'input' })).toHaveFocus();
+        expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+        userEvent.type(screen.getByLabelText('test label', { selector: 'input' }), 'Choi', { skipClick: true });
+        expect(screen.queryByRole('listbox')).toBeInTheDocument();
+      });
 
-    it('does not open when read-only', () => {
-      render(<CustomFieldInputMultipleChoice {...requiredProps} readOnly />);
-      userEvent.click(screen.getByLabelText('test label'));
-      expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+      it('closes on escape key', () => {
+        render(<CustomFieldInputMultipleChoice {...requiredProps} />);
+        userEvent.click(screen.getByText('test label'));
+        expect(screen.queryByRole('listbox')).toBeInTheDocument();
+        fireEvent.keyDown(document.activeElement, { key: 'Escape' });
+        expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+      });
+
+      it('does not open when there are no choices', () => {
+        render(<CustomFieldInputMultipleChoice {...requiredProps} choices={[]} />);
+        userEvent.click(screen.getByText('test label'));
+        expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+      });
+
+      it('does not open when read-only', () => {
+        render(<CustomFieldInputMultipleChoice {...requiredProps} readOnly />);
+        userEvent.click(screen.getByText('test label'));
+        expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+      });
     });
   });
 
@@ -64,9 +106,33 @@ describe('<CustomFieldInputMultipleChoice>', () => {
       expect(screen.getByText('Choice 1')).toBeInTheDocument();
       expect(screen.getByText('Choice 2')).toBeInTheDocument();
 
-      userEvent.click(screen.getAllByRole('button')[0]);
+      userEvent.click(screen.getByRole('button', { name: 'Remove Choice 1' }));
       expect(screen.queryByText('Choice 1')).not.toBeInTheDocument();
       expect(screen.getByText('Choice 2')).toBeInTheDocument();
+    });
+
+    it('does not expand the popup', () => {
+      render((<CustomFieldInputMultipleChoice
+        {...requiredProps}
+        value={requiredProps.choices}
+      />));
+
+      expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+      userEvent.click(screen.getByRole('button', { name: 'Remove Choice 1' }));
+      expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+    });
+
+    it('removes all choices when pressing the clear button', () => {
+      render((<CustomFieldInputMultipleChoice
+        {...requiredProps}
+        value={requiredProps.choices}
+      />));
+
+      expect(screen.getByText('Choice 1')).toBeInTheDocument();
+      expect(screen.getByText('Choice 2')).toBeInTheDocument();
+      userEvent.click(screen.getByRole('button', { name: 'Remove all selected choices on test label' }));
+      expect(screen.queryByText('Choice 1')).not.toBeInTheDocument();
+      expect(screen.queryByText('Choice 2')).not.toBeInTheDocument();
     });
   });
 
@@ -81,7 +147,7 @@ describe('<CustomFieldInputMultipleChoice>', () => {
       />));
 
       expect(screen.queryByRole('option', { name: 'Answer to everything' })).not.toBeInTheDocument();
-      userEvent.click(screen.getByLabelText('test label'));
+      userEvent.click(screen.getByText('test label'));
       expect(screen.getByRole('option', { name: 'Answer to everything' })).toBeInTheDocument();
     });
 
@@ -96,8 +162,36 @@ describe('<CustomFieldInputMultipleChoice>', () => {
         value={choices}
       />));
 
-      userEvent.click(screen.getByLabelText('test label'));
+      userEvent.click(screen.getByText('test label'));
       expect(screen.queryByRole('option', { name: 'Answer to everything' })).not.toBeInTheDocument();
+    });
+  });
+
+  describe('errorText API', () => {
+    it('can bet set', () => {
+      render(<CustomFieldInputMultipleChoice {...requiredProps} errorText="This is an error message." />);
+      expect(screen.queryByText('This is an error message.')).toBeInTheDocument();
+    });
+
+    it('renders 1 icon', () => {
+      render(<CustomFieldInputMultipleChoice {...requiredProps} />);
+      expect(screen.queryAllByRole('img')).toHaveLength(1);
+    });
+
+    it('renders 2 icons', () => {
+      render(<CustomFieldInputMultipleChoice {...requiredProps} errorText="This is an error message." />);
+      expect(screen.queryAllByRole('img')).toHaveLength(2);
+    });
+
+    it('renders 2 icons with selected choices', () => {
+      render((
+        <CustomFieldInputMultipleChoice
+          {...requiredProps}
+          errorText="This is an error message."
+          value={[requiredProps.choices[0]]}
+        />
+      ));
+      expect(screen.queryAllByRole('img')).toHaveLength(2);
     });
   });
 
@@ -118,7 +212,7 @@ describe('<CustomFieldInputMultipleChoice>', () => {
     it('can be set', () => {
       render((<CustomFieldInputMultipleChoice {...requiredProps} label="Unique label" />));
 
-      expect(screen.getByLabelText('Unique label')).toBeInTheDocument();
+      expect(screen.getByLabelText('Unique label', { selector: '[role="grid"]' })).toBeInTheDocument();
     });
   });
 
@@ -140,27 +234,27 @@ describe('<CustomFieldInputMultipleChoice>', () => {
         value={[requiredProps.choices[0]]}
       />));
 
-      expect(screen.getByRole('button')).toBeInTheDocument();
+      expect(screen.getByRole('button', { name: 'Remove Choice 1' })).toBeInTheDocument();
     });
   });
 
   describe('selecting a choice', () => {
     it('removes the choice from the list of choices', () => {
       render(<CustomFieldInputMultipleChoice {...requiredProps} />);
-      userEvent.click(screen.getByLabelText('test label'));
+      userEvent.click(screen.getByText('test label'));
       expect(screen.queryByRole('option', { name: 'Choice 1' })).toBeInTheDocument();
       userEvent.click(screen.getByRole('option', { name: 'Choice 1' }));
-      userEvent.click(screen.getByLabelText('test label'));
+      userEvent.click(screen.getByText('test label'));
       expect(screen.queryByRole('option', { name: 'Choice 1' })).not.toBeInTheDocument();
     });
 
     it('adds the choice to the selected choices', () => {
       render(<CustomFieldInputMultipleChoice {...requiredProps} />);
-      userEvent.click(screen.getByLabelText('test label'));
+      userEvent.click(screen.getByText('test label'));
       expect(screen.queryByRole('gridcell', { name: 'Choice 1' })).not.toBeInTheDocument();
       userEvent.click(screen.queryByRole('option', { name: 'Choice 1' }));
       expect(screen.queryByRole('option', { name: 'Choice 1' })).not.toBeInTheDocument();
-      userEvent.click(screen.getByLabelText('test label'));
+      userEvent.click(screen.getByText('test label'));
       expect(screen.queryByRole('gridcell', { name: 'Choice 1' })).toBeInTheDocument();
     });
   });