+++ src/components/custom-field-input-number/custom-field-input-number.test.jsx
@@ -28,7 +28,7 @@ describe('CustomFieldInputNumber', () => {
     });
 
     it('can be enabled', () => {
-      const { container } = render(<TestComponent />);
+      render(<TestComponent />);
       expect(screen.getByLabelText('Test label')).not.toBeDisabled();
     });
   });