+++ src/components/custom-field-input-single-choice/custom-field-input-single-choice.jsx
@@ -8,25 +8,44 @@ import React, {
 import PropTypes from 'prop-types';
 import AbstractCustomField from '../__internal__/abstract-custom-field.jsx';
 import Icon from '../icon/icon.jsx';
+import iconClear from '../../svgs/icon-clear-small.svg';
 import iconCaretDown from '../../svgs/icon-caret-down.svg';
 import iconCaretDownDisabled from '../../svgs/icon-caret-down-disabled.svg';
 import styles from './custom-field-input-single-choice.css';
 import Listbox from '../listbox/listbox.jsx';
 import ListOption from '../list-option/list-option.jsx';
+import useValidation from '../../hooks/use-validation.jsx';
 
 const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleChoice(props, ref) {
   const [didMount, setDidMount] = useState(false);
   const [showOptions, setShowOptions] = useState(false);
   const [value, setValue] = useState(props.value);
   const [searchValue, setSearchValue] = useState(undefined);
-
   const inputRef = useRef();
+
+  const validationMessage = useValidation(props.readOnly, props.errorText, inputRef, false);
   const refs = props.choices.map(() => useRef());
   const caretIcon = (<Icon
     className={styles['input-icon']}
     name={props.readOnly ? iconCaretDownDisabled.id : iconCaretDown.id}
     fill="skip"
   />);
+
+  const clear = () => {
+    setValue(undefined);
+    setSearchValue(undefined);
+  };
+
+  const clearIcon = () => {
+    if (!props.readOnly && (value || searchValue)) {
+      return (<Icon
+        name={iconClear.id}
+        onClick={clear}
+      />);
+    }
+
+    return undefined;
+  };
 
   function onClick() {
     if (!props.readOnly) setShowOptions(true);
@@ -118,6 +137,7 @@ const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleC
     <div className={styles.container}>
       <AbstractCustomField
         icon={caretIcon}
+        clear={clearIcon()}
         id={props.id}
         label={props.label}
         onChange={onSearchChange}
@@ -127,6 +147,7 @@ const CustomFieldInputSingleChoice = forwardRef(function CustomFieldInputSingleC
         readOnly={props.readOnly}
         inputRef={inputRef}
         required={props.required}
+        errorText={validationMessage}
         value={searchValue || (value ? value.label : '')}
       />
       { showOptions && (
@@ -157,6 +178,7 @@ CustomFieldInputSingleChoice.propTypes = {
   readOnly: PropTypes.bool,
   required: PropTypes.bool,
   value: ChoiceType,
+  errorText: PropTypes.string,
 };
 
 CustomFieldInputSingleChoice.defaultProps = {
@@ -165,6 +187,7 @@ CustomFieldInputSingleChoice.defaultProps = {
   readOnly: false,
   required: false,
   value: undefined,
+  errorText: undefined,
 };
 
 export default CustomFieldInputSingleChoice;